// src/app/modules/estudiante/models/itemmatricula.model.ts
export interface ItemMatricula{
    id_matricula: number;
    estudiante_matricula: number;
    fecha_matricula: string;
    ciclo_matricula: string;
    id_itemMatricula?: number;
    matricula_pagoMatricula?: number;
    item_pagoMatricula?: string;
    cantidad_pagoMatricula?: number;
    subtotal_pagoMatricula?: number;
    descuento_pagoMatricula?: number;
    total_pagoMatricula?: number;
}

  